backend/requirements.txt
fastapi
httpx
langchain
langchain-google-genai
python-dotenv
uvicorn
